+++ packages/web3-utils/src/index.js
@@ -23,11 +23,9 @@
 
 
 var _ = require('underscore');
-var BN = require('bn.js');
 var ethjsUnit = require('ethjs-unit');
-var numberToBN = require('number-to-bn');
-var utf8 = require('utf8');
-var jsSha3 = require("js-sha3");
+var utils = require('./utils.js');
+var soliditySha3 = require('./soliditySha3.js');
 
 
 /**
@@ -94,383 +92,8 @@ var _jsonInterfaceMethodToString = function (json) {
     return json.name + '(' + typeName + ')';
 };
 
-/**
- * Convert a hex string to a byte array
- *
- * Note: Implementation from crypto-js
- *
- * @method hexToBytes
- * @param {string} hex
- * @return {Array} the byte array
- */
-var hexToBytes = function(hex) {
-    hex = hex.toString(16);
 
-    if (!isHex(hex)) {
-        throw new Error('Given value "'+ hex +'" is not a valid hex string.');
-    }
 
-    hex = hex.replace(/^0x/i,'');
-
-    for (var bytes = [], c = 0; c < hex.length; c += 2)
-        bytes.push(parseInt(hex.substr(c, 2), 16));
-    return bytes;
-};
-
-
-/**
- * Convert a byte array to a hex string
- *
- * Note: Implementation from crypto-js
- *
- * @method bytesToHex
- * @param {Array} bytes
- * @return {String} the hex string
- */
-var bytesToHex = function(bytes) {
-    for (var hex = [], i = 0; i < bytes.length; i++) {
-        /* jshint ignore:start */
-        hex.push((bytes[i] >>> 4).toString(16));
-        hex.push((bytes[i] & 0xF).toString(16));
-        /* jshint ignore:end */
-    }
-    return '0x'+ hex.join("");
-};
-
-
-/**
- * Hashes values to a sha3 hash using keccak 256
- *
- * To hash a HEX string the hex must have 0x in front.
- *
- * @method sha3
- * @return {String} the sha3 string
- */
-var SHA3_NULL_S = '0xc5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470';
-
-var sha3 = function (value) {
-    if (isHex(value) && /^0x/i.test((value).toString())) {
-        value = hexToBytes(value);
-    }
-
-    var returnValue = '0x'+ jsSha3.keccak_256(value); // jshint ignore:line
-
-    if(returnValue === SHA3_NULL_S) {
-        return null;
-    } else {
-        return returnValue;
-    }
-};
-// expose the under the hood keccak256
-sha3.jsSha3 = jsSha3;
-
-
-
-var _elementaryName = function (name) {
-    /*jshint maxcomplexity:false */
-
-    if (name.startsWith('int[')) {
-        return 'int256' + name.slice(3);
-    } else if (name === 'int') {
-        return 'int256';
-    } else if (name.startsWith('uint[')) {
-        return 'uint256' + name.slice(4);
-    } else if (name === 'uint') {
-        return 'uint256';
-    } else if (name.startsWith('fixed[')) {
-        return 'fixed128x128' + name.slice(5);
-    } else if (name === 'fixed') {
-        return 'fixed128x128';
-    } else if (name.startsWith('ufixed[')) {
-        return 'ufixed128x128' + name.slice(6);
-    } else if (name === 'ufixed') {
-        return 'ufixed128x128';
-    }
-    return name;
-};
-
-// Parse N from type<N>
-var _parseTypeN = function (type) {
-    return parseInt(/^\D+(\d+).*$/.exec(type)[1], 10);
-};
-
-// Parse N from type[<N>]
-var _parseTypeNArray = function (type) {
-    var arraySize = /^\D+\d*\[(\d+)\]$/.exec(type);
-    return arraySize ? parseInt(arraySize[1], 10) : null;
-};
-
-var _parseNumber = function (arg) {
-    var type = typeof arg;
-    if (type === 'string') {
-        if (isHex(arg)) {
-            return new BN(arg.replace(/0x/i,''), 16);
-        } else {
-            return new BN(arg, 10);
-        }
-    } else if (type === 'number') {
-        return new BN(arg);
-    } else if (isBigNumber(arg)) {
-        return new BN(arg.toString(10));
-    } else if (isBN(arg)) {
-        return arg;
-    } else {
-        throw new Error(arg +' is not a number');
-    }
-};
-
-var _solidityPack = function (type, value, arraySize) {
-    /*jshint maxcomplexity:false */
-
-    var size, num;
-    type = _elementaryName(type);
-
-
-    if (type === 'bytes') {
-        return value;
-    } else if (type === 'string') {
-        return utf8ToHex(value);
-    } else if (type === 'bool') {
-        return value ? '01' : '00';
-    } else if (type.startsWith('address')) {
-        if(arraySize) {
-            size = 64;
-        } else {
-            size = 40;
-        }
-
-        return leftPad(value.toLowerCase(), size);
-    } else if (type.startsWith('bytes')) {
-        size = _parseTypeN(type);
-
-        // must be 32 byte slices when in an array
-        if(arraySize) {
-            size = 32;
-        }
-
-        if (size < 1 || size > 32) {
-            throw new Error('Invalid bytes<N> width: ' + size);
-        }
-
-        return rightPad(value, size * 2);
-    } else if (type.startsWith('uint')) {
-        size = _parseTypeN(type);
-
-        if ((size % 8) || (size < 8) || (size > 256)) {
-            throw new Error('Invalid uint<N> width: ' + size);
-        }
-
-        num = _parseNumber(value);
-        if (num.bitLength() > size) {
-            throw new Error('Supplied uint exceeds width: ' + size + ' vs ' + num.bitLength());
-        }
-
-        if(num.lt(new BN(0))) {
-            throw new Error('Supplied uint '+ num.toString() +' is negative');
-        }
-
-        return size ? leftPad(num.toString('hex'), size/8 * 2) : num;
-    } else if (type.startsWith('int')) {
-
-        size = _parseTypeN(type);
-        if ((size % 8) || (size < 8) || (size > 256)) {
-            throw new Error('Invalid int<N> width: ' + size);
-        }
-
-        num = _parseNumber(value);
-        if (num.bitLength() > size) {
-            throw new Error('Supplied int exceeds width: ' + size + ' vs ' + num.bitLength());
-        }
-
-        if(num.lt(new BN(0))) {
-            return num.toTwos(size).toString('hex');
-        } else {
-            return size ? leftPad(num.toString('hex'), size/8 * 2) : num;
-        }
-
-    } else {
-        // FIXME: support all other types
-        throw new Error('Unsupported or invalid type: ' + type);
-    }
-};
-
-
-var _processSoliditySha3Args = function (arg) {
-    /*jshint maxcomplexity:false */
-    
-    var type, value = '';
-    var hexArg, arraySize;
-
-    // if type is given
-    if (_.isObject(arg) && (arg.hasOwnProperty('v') || arg.hasOwnProperty('t') || arg.hasOwnProperty('value') || arg.hasOwnProperty('type'))) {
-        type = arg.t || arg.type;
-        value = arg.v || arg.value;
-
-        // otherwise try to guess the type
-    } else {
-        type = toHex(arg, true);
-        value = toHex(arg);
-
-        if (!type.startsWith('int') && !type.startsWith('uint')) {
-            type = 'bytes';
-        }
-    }
-
-    if ((type.startsWith('int') || type.startsWith('uint')) &&  typeof value === 'string' && !/^(-)?0x/i.test(value)) {
-        value = new BN(value);
-    }
-
-    // get the array size
-    if(_.isArray(value)) {
-        arraySize = _parseTypeNArray(type);
-        if(arraySize && value.length !== arraySize) {
-            throw new Error(type +' is not matching the given array '+ JSON.stringify(value));
-        } else {
-            arraySize = value.length;
-        }
-    }
-
-
-    if (_.isArray(value)) {
-        hexArg = value.map(function (val) {
-            return _solidityPack(type, val, arraySize).toString('hex').replace('0x','');//abi.encodeParam(arg.type, arg.value);
-        });
-        return hexArg.join('');
-    } else {
-        hexArg = _solidityPack(type, value, arraySize);//abi.encodeParam(arg.type, arg.value);
-        return hexArg.toString('hex').replace('0x','');
-    }
-
-};
-
-/**
- * Hashes solidity values to a sha3 hash using keccak 256
- *
- * @method soliditySha3
- * @return {Object} the sha3
- */
-var soliditySha3 = function () {
-    /*jshint maxcomplexity:false */
-
-    var args = Array.prototype.slice.call(arguments);
-
-    var hexArgs = _.map(args, _processSoliditySha3Args);
-
-    // console.log(args, hexArgs);
-    // console.log('0x'+ hexArgs.join(''));
-
-    return sha3('0x'+ hexArgs.join(''));
-};
-
-
-/**
- * Should be called to pad string to expected length
- *
- * @method leftPad
- * @param {String} string to be padded
- * @param {Number} chars that result string should have
- * @param {String} sign, by default 0
- * @returns {String} right aligned string
- */
-var leftPad = function (string, chars, sign) {
-    var hasPrefix = /^0x/i.test(string) || typeof string === 'number';
-    string = string.toString(16).replace(/^0x/i,'');
-
-    var padding = (chars - string.length + 1 >= 0) ? chars - string.length + 1 : 0;
-
-    return (hasPrefix ? '0x' : '') + new Array(padding).join(sign ? sign : "0") + string;
-};
-
-/**
- * Should be called to pad string to expected length
- *
- * @method rightPad
- * @param {String} string to be padded
- * @param {Number} chars that result string should have
- * @param {String} sign, by default 0
- * @returns {String} right aligned string
- */
-var rightPad = function (string, chars, sign) {
-    var hasPrefix = /^0x/i.test(string) || typeof string === 'number';
-    string = string.toString(16).replace(/^0x/i,'');
-
-    var padding = (chars - string.length + 1 >= 0) ? chars - string.length + 1 : 0;
-
-    return (hasPrefix ? '0x' : '') + string + (new Array(padding).join(sign ? sign : "0"));
-};
-
-/**
- * Check if string is HEX
- *
- * @method isHex
- * @param {String} hex to be checked
- * @returns {Boolean}
- */
-var isHex = function (hex) {
-    return ((_.isString(hex) || _.isNumber(hex)) && /^(-)?0x[0-9a-f]+$/i.test(hex));
-};
-
-
-/**
- * Should be called to get utf8 from it's hex representation
- *
- * @method hexToUtf8
- * @param {String} hex
- * @returns {String} ascii string representation of hex value
- */
-var hexToUtf8 = function(hex) {
-    if (!isHex(hex))
-        throw new Error('The parameter must be a valid HEX string.');
-
-    var str = "";
-    var code = 0;
-    hex = hex.replace(/^0x/i,'');
-
-    // remove 00 padding from either side
-    hex = hex.replace(/^(?:00)*/,'');
-    hex = hex.split("").reverse().join("");
-    hex = hex.replace(/^(?:00)*/,'');
-    hex = hex.split("").reverse().join("");
-
-    var l = hex.length;
-
-    for (var i=0; i < l; i+=2) {
-        code = parseInt(hex.substr(i, 2), 16);
-        // if (code !== 0) {
-        str += String.fromCharCode(code);
-        // }
-    }
-
-    return utf8.decode(str);
-};
-
-/**
- * Should be called to get hex representation (prefixed by 0x) of utf8 string
- *
- * @method utf8ToHex
- * @param {String} str
- * @returns {String} hex representation of input string
- */
-var utf8ToHex = function(str) {
-    str = utf8.encode(str);
-    var hex = "";
-
-    // remove \u0000 padding from either side
-    str = str.replace(/^(?:\u0000)*/,'');
-    str = str.split("").reverse().join("");
-    str = str.replace(/^(?:\u0000)*/,'');
-    str = str.split("").reverse().join("");
-
-    for(var i = 0; i < str.length; i++) {
-        var code = str.charCodeAt(i);
-        // if (code !== 0) {
-            var n = code.toString(16);
-            hex += n.length < 2 ? '0' + n : n;
-        // }
-    }
-
-    return "0x" + hex;
-};
 
 /**
  * Should be called to get ascii from it's hex representation
@@ -480,7 +103,7 @@ var utf8ToHex = function(str) {
  * @returns {String} ascii string representation of hex value
  */
 var hexToAscii = function(hex) {
-    if (!isHex(hex))
+    if (!utils.isHex(hex))
         throw new Error('The parameter must be a valid HEX string.');
 
     var str = "";
@@ -515,85 +138,6 @@ var asciiToHex = function(str) {
 };
 
 
-/**
- * Converts value to it's number representation
- *
- * @method hexToNumber
- * @param {String|Number|BN} value
- * @return {String}
- */
-var hexToNumber = function (value) {
-    if (!value) return value;
-
-    return toBN(value).toNumber();
-};
-
-/**
- * Converts value to it's decimal representation in string
- *
- * @method hexToNumberString
- * @param {String|Number|BN} value
- * @return {String}
- */
-var hexToNumberString = function (value) {
-    if (!value) return value;
-
-    return toBN(value).toString(10);
-};
-
-/**
- * Converts value to it's hex representation
- *
- * @method numberToHex
- * @param {String|Number|BN} value
- * @return {String}
- */
-var numberToHex = function (value) {
-    var number = toBN(value);
-    var result = number.toString(16);
-
-    return number.lt(new BN(0)) ? '-0x' + result.substr(1) : '0x' + result;
-};
-
-/**
- * Auto converts any given value into it's hex representation.
- *
- * And even stringifys objects before.
- *
- * @method toHex
- * @param {String|Number|BN|Object} value
- * @param {Boolean} returnType
- * @return {String}
- */
-var toHex = function (value, returnType) {
-    /*jshint maxcomplexity: false */
-
-    if (isAddress(value)) {
-        return returnType ? 'address' : '0x'+ value.toLowerCase().replace(/^0x/i,'');
-    }
-
-    if (_.isBoolean(value)) {
-        return returnType ? 'bool' : value ? '0x01' : '0x00';
-    }
-
-
-    if (_.isObject(value) && !isBigNumber(value) && !isBN(value)) {
-        return returnType ? 'string' : utf8ToHex(JSON.stringify(value));
-    }
-
-    // if its a negative number, pass it through numberToHex
-    if (_.isString(value)) {
-        if (value.indexOf('-0x') === 0 || value.indexOf('-0X') === 0) {
-            return returnType ? 'int256' : numberToHex(value);
-        } else if(value.indexOf('0x') === 0 || value.indexOf('0X') === 0) {
-            return returnType ? 'bytes' : value;
-        } else if (!isFinite(value)) {
-            return returnType ? 'string' : utf8ToHex(value);
-        }
-    }
-
-    return returnType ? (value < 0 ? 'int256' : 'uint256') : numberToHex(value);
-};
 
 /**
  * Returns value of unit in Wei
@@ -635,7 +179,7 @@ var getUnitValue = function (unit) {
 var fromWei = function(number, unit) {
     unit = getUnitValue(unit);
 
-    return isBN(number) ? ethjsUnit.fromWei(number, unit) : ethjsUnit.fromWei(number, unit).toString(10);
+    return utils.isBN(number) ? ethjsUnit.fromWei(number, unit) : ethjsUnit.fromWei(number, unit).toString(10);
 };
 
 /**
@@ -663,66 +207,9 @@ var fromWei = function(number, unit) {
 var toWei = function(number, unit) {
     unit = getUnitValue(unit);
 
-    return isBN(number) ? ethjsUnit.toWei(number, unit) : ethjsUnit.toWei(number, unit).toString(10);
+    return utils.isBN(number) ? ethjsUnit.toWei(number, unit) : ethjsUnit.toWei(number, unit).toString(10);
 };
 
-/**
- * Takes an input and transforms it into an BN
- *
- * @method toBN
- * @param {Number|String|BN} number, string, HEX string or BN
- * @return {BN} BN
- */
-var toBN = function(number){
-    try {
-        return numberToBN.apply(null, arguments);
-    } catch(e) {
-        throw new Error(e + ' Given value: "'+ number +'"');
-    }
-};
-
-/**
- * Checks if the given string is an address
- *
- * @method isAddress
- * @param {String} address the given HEX address
- * @return {Boolean}
- */
-var isAddress = function (address) {
-    // check if it has the basic requirements of an address
-    if (!/^(0x)?[0-9a-f]{40}$/i.test(address)) {
-        return false;
-    // If it's ALL lowercase or ALL upppercase
-    } else if (/^(0x|0X)?[0-9a-f]{40}$/.test(address) || /^(0x|0X)?[0-9A-F]{40}$/.test(address)) {
-        return true;
-    // Otherwise check each case
-    } else {
-        return checkAddressChecksum(address);
-    }
-};
-
-
-
-/**
- * Checks if the given string is a checksummed address
- *
- * @method checkAddressChecksum
- * @param {String} address the given HEX address
- * @return {Boolean}
- */
-var checkAddressChecksum = function (address) {
-    // Check each case
-    address = address.replace(/^0x/i,'');
-    var addressHash = sha3(address.toLowerCase()).replace(/^0x/i,'');
-
-    for (var i = 0; i < 40; i++ ) {
-        // the nth letter should be uppercase if the nth digit of casemap is 1
-        if ((parseInt(addressHash[i], 16) > 7 && address[i].toUpperCase() !== address[i]) || (parseInt(addressHash[i], 16) <= 7 && address[i].toLowerCase() !== address[i])) {
-            return false;
-        }
-    }
-    return true;
-};
 
 
 
@@ -742,7 +229,7 @@ var toChecksumAddress = function (address) {
 
 
     address = address.toLowerCase().replace(/^0x/i,'');
-    var addressHash = sha3(address).replace(/^0x/i,'');
+    var addressHash = utils.sha3(address).replace(/^0x/i,'');
     var checksumAddress = '0x';
 
     for (var i = 0; i < address.length; i++ ) {
@@ -757,29 +244,6 @@ var toChecksumAddress = function (address) {
 };
 
 
-/**
- * Returns true if object is BN, otherwise false
- *
- * @method isBN
- * @param {Object} object
- * @return {Boolean}
- */
-var isBN = function (object) {
-    return object instanceof BN ||
-        (object && object.constructor && object.constructor.name === 'BN');
-};
-
-/**
- * Returns true if object is BigNumber, otherwise false
- *
- * @method isBigNumber
- * @param {Object} object
- * @return {Boolean}
- */
-var isBigNumber = function (object) {
-    return object && object.constructor && object.constructor.name === 'BigNumber';
-};
-
 
 module.exports = {
     _fireError: _fireError,
@@ -787,37 +251,37 @@ module.exports = {
     // extractDisplayName: extractDisplayName,
     // extractTypeName: extractTypeName,
     _: _,
-    BN: BN,
-    isBN: isBN,
-    isBigNumber: isBigNumber,
-    isHex: isHex,
-    sha3: sha3,
-    keccak256: sha3,
+    BN: utils.BN,
+    isBN: utils.isBN,
+    isBigNumber: utils.isBigNumber,
+    isHex: utils.isHex,
+    sha3: utils.sha3,
+    keccak256: utils.sha3,
     soliditySha3: soliditySha3,
-    isAddress: isAddress,
-    checkAddressChecksum: checkAddressChecksum,
+    isAddress: utils.isAddress,
+    checkAddressChecksum: utils.checkAddressChecksum,
     toChecksumAddress: toChecksumAddress,
-    toHex: toHex,
-    toBN: toBN,
+    toHex: utils.toHex,
+    toBN: utils.toBN,
 
-    bytesToHex: bytesToHex,
-    hexToBytes: hexToBytes,
+    bytesToHex: utils.bytesToHex,
+    hexToBytes: utils.hexToBytes,
 
-    hexToNumberString: hexToNumberString,
+    hexToNumberString: utils.hexToNumberString,
 
-    hexToNumber: hexToNumber,
-    toDecimal: hexToNumber, // alias
+    hexToNumber: utils.hexToNumber,
+    toDecimal: utils.hexToNumber, // alias
 
-    numberToHex: numberToHex,
-    fromDecimal: numberToHex, // alias
+    numberToHex: utils.numberToHex,
+    fromDecimal: utils.numberToHex, // alias
 
-    hexToUtf8: hexToUtf8,
-    hexToString: hexToUtf8,
-    toUtf8: hexToUtf8,
+    hexToUtf8: utils.hexToUtf8,
+    hexToString: utils.hexToUtf8,
+    toUtf8: utils.hexToUtf8,
 
-    utf8ToHex: utf8ToHex,
-    stringToHex: utf8ToHex,
-    fromUtf8: utf8ToHex,
+    utf8ToHex: utils.utf8ToHex,
+    stringToHex: utils.utf8ToHex,
+    fromUtf8: utils.utf8ToHex,
 
     hexToAscii: hexToAscii,
     toAscii: hexToAscii,
@@ -828,9 +292,9 @@ module.exports = {
     toWei: toWei,
     fromWei: fromWei,
 
-    padLeft: leftPad,
-    leftPad: leftPad,
-    padRight: rightPad,
-    rightPad: rightPad
+    padLeft: utils.leftPad,
+    leftPad: utils.leftPad,
+    padRight: utils.rightPad,
+    rightPad: utils.rightPad
 };
